pyproject.toml
[project]
name = "deepseek-ocr-pdf2md"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "flash-attn @ git+https://github.com/Dao-AILab/flash-attention.git@v2.7.3",
    "tokenizers>=0.21.1",
    "PyMuPDF",
    "img2pdf",
    "einops",
    "easydict",
    "addict",
    "Pillow",
    "numpy",
    "transformers",
    "xformers",
    "vllm",
    "triton-kernels",
    "torch",
    "torchvision",
    "torchaudio",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.6.0",
    "typer>=0.12.3",
]

[project.scripts]
deepseek-pdf2md = "deepseek_pdf2md.cli:app"

[tool.setuptools]
packages = ["deepseek_pdf2md"]

[tool.uv.sources]
vllm = [
  { index = "vllm-nightly" },
]
triton-kernels = { git = "https://github.com/triton-lang/triton.git", rev = "v3.5.0", subdirectory = "python/triton_kernels" }
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "vllm-nightly"
url = "https://wheels.vllm.ai/nightly"
explicit = true

[tool.uv]
no-build-isolation-package = ["vllm", "torch", "flash-attn"]
override-dependencies = [
    "xformers @ git+https://github.com/facebookresearch/xformers.git@5d4b92a5",
]

[tool.uv.extra-build-dependencies]
vllm = [
    "cmake>=3.26.1",
    "ninja",
    "packaging>=24.2",
    "setuptools>=77.0.3,<80.0.0",
    "setuptools-scm>=8.0",
    "wheel",
    "jinja2",
]
flash-attn = [
    "torch",
]
